+++ serde_valid/tests/deprecated_rule_test.rs
